Can now run a flow using DHF 5.2.0 (#24)
* Updating the AWS quickstart document.

* Updating the AWS quickstart document. (#19)

* Can now run a flow using DHF 5.2.0 (#21)

The key changes are in MarkLogicSinkTask and RunFlowWriteBatchListener. See the comments in those classes for the changes.

There are 3 new properties for using this - a DHF flow name, and optional set of steps, and an option for logging each flow response. If the flow name is set, then that flow will be run. Otherwise, only the regular ingestion happens. 

I had to make one plumbing change - DatabaseClientCreator is now DatabaseClientConfigBuilder, as I needed to reuse the DatabaseClientConfig object. That's a small plumbing change though that could be made independently of this change.

Co-authored-by: Rob Rudin <rob.rudin@marklogic.com>

* Updating the AWS quickstart document. (#23)

Co-authored-by: rjrudin <rjrudin@gmail.com>
Co-authored-by: Rob Rudin <rob.rudin@marklogic.com>