Develop (#29)
* Incremented the minor version in the develop branch for new work.

* update to CloudFormation templates and instructions (#25)

* Can now run a flow using DHF 5.2.0

The key changes are in MarkLogicSinkTask and RunFlowWriteBatchListener. See the comments in those classes for the changes.

There are 3 new properties for using this - a DHF flow name, and optional set of steps, and an option for logging each flow response. If the flow name is set, then that flow will be run. Otherwise, only the regular ingestion happens. 

I had to make one plumbing change - DatabaseClientCreator is now DatabaseClientConfigBuilder, as I needed to reuse the DatabaseClientConfig object. That's a small plumbing change though that could be made independently of this change.

Co-authored-by: Rob Rudin <rob.rudin@marklogic.com>

* Updating the AWS quickstart document.
* Update AWS-CloudFormation/cloudFormationTemplates/* with newer AMI, added some steps to CloudFormation-QuickStart.md

Co-authored-by: Stuart Moorhouse <Stuart.Moorhouse@marklogic.com>

* Removing some erroneous comments in the example connect-standalone properties file. (#28)

Co-authored-by: sMoorhouse <stuartmoorhouse@rocketmail.com>
Co-authored-by: Stuart Moorhouse <Stuart.Moorhouse@marklogic.com>