Updating README with a user guide
Rearranged some code in DefaultDatabaseClientConfigBuilder while I was trying to understand the custom SSL support. 

Removed the 3 PDF docs. 1.2.2 and 1.3.0 were identical, and I believe the README does a better job of addressing how to configure custom SSL stuff. Created #72 to capture tests and docs for JSON/Avro converters. And attached the 1.2.3 doc to #69 for documenting ID strategies. 

Removed CHANGELOG in favor of maintaining release notes in each GitHub release, which we've already been doing. 

Simplified the marklogic-connect-standalone.properties files to better describe what these files are for and why they would need to be modified. The main things to modify in them are the key/value converters, which will be addressed by #72.