Reworked Bulk DS approach to avoid reading from modules database
This avoids requiring the ML user to have an eval privilege. 

The connector user now just needs to configure the endpoint URI and an optional batch size. The Bulk DS endpoint must still have an API declaration in the modules database, as that's required by Bulk DS. But the connector now builds an API declaration in memory based on the user's inputs and the known set of params, thus avoiding having to read from the modules database.