Switching from RowBatcher to RowManager
Consensus was that the downsides of RowBatcher - can only use fromView; errors happening in the background; confusion over what "batchSize" refers to - were not worth the potential benefit in performance. That's because once RowBatcher is noticeably faster than RowManager, the user is likely pulling back enough rows to run out of memory / Java heap space due to a massive List. 

This leaves all the RowBatcher support in place for now. Main addition is a `PlanInvoker` that does what `RowBatcherBuilder` effectively did, which is to choose the correct handle, invoke a plan, and convert the results into source records. 

Once this is merged, I'll submit a PR to remove all the RB support and rename the task to RowManagerSourceTask. Having that all in one commit/PR will make it easy to pull from if we decide to add RB support back in the future.