DEVEXP-127: Using ML server timestamp for key strategy
Also adjusted the `KeyGenerator` interface to not require a row number so that the plan invokers don't have to manage that; the timestamp generator can easily manage that itself. In addition, a new generator is constructed each time one or more rows are returned. 

Added a README section for key strategy as well.