Replaced key strategy with key column
Based on feedback from Confluent colleagues. 

With only one strategy, the approach for extracting a key value is located in each `PlanInvoker`, as the approach depends on the type of data returned by MarkLogic. If/when we gain a second strategy (other than "return null"), we can revisit this design. 

Did a little cleanup by replacing remaining "RowBatcher" occurrences.