DEVEXP-652 Bumped to Java Client 6.3.0 to address vulnerability
Addresses the vulnerability in okio, which is fixed in Java Client 6.3.0. 

Had to bump Gradle to 8 to allow for uber jar to be built, now that Jackson 2.15 is used by the Java Client (we did the same thing for our Spark connector). 

Everything in the CONTRIB ran fine, no issues encountered.